Finish setup
Install webpack merge.
Separate the webpack config
files into common,
development and
production with each file
handling the config its
environment.

Configure the asset modules
in webpack common to load
fonts. Also add clean: true
flag to output in webpack
common so that it removes
any file in the output
directory (dist)
before generating new files
and ensure that it is the
latest build output that
is kept in the there.

Edit the start and build
scripts in package.json
to run the
development and production
environments from their
files.

The template is thus ready.