Setup eslint and prettier
Install eslint and the
eslint-config-airbnb-base
community plugin.

Install the
eslint-config-prettier plugin,
config that turns off all
of eslint's configuration
for things that prettier
already handles as
eslint-config-airbnb-base turns
on some stylistic rules that may
clash with Prettier's formatting.
Add prettier as the last element
on the array of the .eslintrc
extends key so that the
plugin can work as expected.

Add a .prettierrc
For prettier customizations instead
of customizing in the package.json

Add a .prettierignore file to
let prettier know what files and
formats to ignore.

Update the name in package.json
to webpack-config.